+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [3.0.6] - Unreleased
 ### Added
 - New `idle` event, triggered when the `build` or `update` process has finished or stopped.
+- New flag `--cms` for `deno task serve` to run the new experimental version of LumeCMS.
 
 ### Changed
 - `esbuild` plugin: use `jsr:@deno/loader` official package instead of `@luca/esbuild-deno-loader`.